fix(Scrolling): Terminating Responder on Scroll
This (hack) causes a Touchable to enter a terminated state when the browser window scrolls. It does not fix the React responder system to properly terminate the responder so events will continue to flow into the Touchable which must be ignored.

This fixes [an issue](/necolas/react-native-web/issues/722) where `onPress` events are still being sent when the user scrolls outside of a React element (like on a normal webpage).